- updated some file names to make 'local' (spiNNlink) and 'remote' (spif) consistent across test code and documentation.
