+++ doc/spif_config/spif_bd.fig
@@ -76,8 +76,6 @@ Single
 	 9450 2385 10350 2385
 2 2 0 1 0 11 50 -1 20 0.000 0 0 -1 0 0 5
 	 7875 2025 9450 2025 9450 2700 7875 2700 7875 2025
-2 2 0 1 0 11 50 -1 20 0.000 0 0 -1 0 0 5
-	 10350 1800 11925 1800 11925 2925 10350 2925 10350 1800
 2 1 0 3 0 7 45 -1 -1 0.000 0 0 -1 1 0 2
 	1 1 3.00 180.00 360.00
 	 13725 4185 14625 4185
@@ -86,27 +84,14 @@ Single
 2 1 0 3 0 7 45 -1 -1 0.000 0 0 -1 1 0 2
 	1 1 3.00 180.00 360.00
 	 13725 2385 14625 2385
-2 2 0 1 0 20 50 -1 20 0.000 0 0 -1 0 0 5
-	 11925 1800 13725 1800 13725 2925 11925 2925 11925 1800
 2 1 0 3 0 7 45 -1 -1 0.000 0 0 -1 1 0 2
 	1 1 3.00 180.00 360.00
 	 15750 3285 16650 3285
 2 2 0 1 0 20 50 -1 20 0.000 0 0 -1 0 0 5
 	 16650 2250 18450 2250 18450 4275 16650 4275 16650 2250
-2 2 0 1 0 7 60 -1 15 0.000 0 0 -1 0 0 5
-	 14715 5940 16515 5940 16515 6840 14715 6840 14715 5940
-2 2 0 3 13 14 50 -1 30 0.000 0 0 -1 0 0 5
-	 14625 5850 16425 5850 16425 6750 14625 6750 14625 5850
 2 1 0 3 13 7 45 -1 -1 0.000 0 0 -1 1 0 3
 	1 1 3.00 180.00 360.00
-	 14625 6075 12825 6075 12825 4725
-2 1 0 3 13 7 45 -1 -1 0.000 0 0 -1 1 0 3
-	1 1 3.00 180.00 360.00
-	 16425 6075 17550 6075 17550 4275
-2 1 0 3 13 7 45 -1 -1 0.000 0 0 -1 1 1 2
-	1 1 3.00 180.00 360.00
-	1 1 3.00 180.00 360.00
-	 19350 6525 16425 6525
+	 16425 5850 17550 5850 17550 4275
 2 2 0 1 0 7 75 -1 20 0.000 0 0 -1 0 0 5
 	 19935 1125 7200 1125 7200 7875 19935 7875 19935 1125
 2 1 0 3 13 7 45 -1 -1 0.000 0 0 -1 1 1 2
@@ -137,7 +122,26 @@ Single
 	 7200 1125 900 1125 900 7875 7200 7875 7200 1125
 2 1 0 3 13 7 60 -1 -1 0.000 0 0 -1 1 0 3
 	1 1 3.00 180.00 360.00
-	 14625 6300 11160 6300 11160 2925
+	 14625 5625 13050 5625 13050 2925
+2 1 0 3 13 7 60 -1 -1 0.000 0 0 -1 1 0 3
+	1 1 3.00 180.00 360.00
+	 14625 6075 11475 6075 11475 4725
+2 1 0 3 13 7 60 -1 -1 0.000 0 0 -1 1 0 3
+	1 1 3.00 180.00 360.00
+	 14625 5850 12375 5850 12375 4725
+2 1 0 3 13 7 60 -1 -1 0.000 0 0 -1 1 0 3
+	1 1 3.00 180.00 360.00
+	 14625 6300 10800 6300 10800 2925
+2 1 0 3 13 7 45 -1 -1 0.000 0 0 -1 1 1 2
+	1 1 3.00 180.00 360.00
+	1 1 3.00 180.00 360.00
+	 19350 6300 16425 6300
+2 2 0 1 0 11 50 -1 20 0.000 0 0 -1 0 0 5
+	 10350 1800 11925 1800 11925 2925 10350 2925 10350 1800
+2 2 0 1 0 20 50 -1 20 0.000 0 0 -1 0 0 5
+	 11925 1800 13725 1800 13725 2925 11925 2925 11925 1800
+2 2 0 1 -1 14 50 -1 30 0.000 0 0 -1 0 0 5
+	 14625 5400 16425 5400 16425 6750 14625 6750 14625 5400
 4 1 0 40 -1 18 24 0.0000 4 300 1065 8685 4320 dmac\001
 4 1 0 40 -1 18 24 0.0000 4 300 900 11160 4320 filter\001
 4 1 0 40 -1 18 24 0.0000 4 300 1065 8685 2475 dmac\001
@@ -145,10 +149,12 @@ Single
 4 1 31 40 -1 18 24 0.0000 4 315 1485 12825 4320 mapper\001
 4 1 31 40 -1 18 24 0.0000 4 315 1485 12825 2520 mapper\001
 4 1 31 40 -1 18 24 0.0000 4 285 1200 17595 3375 router\001
-4 1 0 50 -1 18 24 0.0000 4 390 1245 15525 6435 config\001
 4 1 0 40 -1 18 24 0.0000 4 300 615 2790 4320 eth\001
 4 1 0 40 -1 18 24 0.0000 4 225 870 5400 6660 core\001
 4 1 0 40 -1 18 24 0.0000 4 300 615 2790 2475 eth\001
 4 1 0 40 -1 18 24 0.0000 4 225 945 5400 3420 mem\001
 4 1 0 50 -1 18 24 0.0000 4 390 2850 2700 7515 spif processor\001
 4 1 0 50 -1 18 24 0.0000 4 390 3240 17910 7515 spif FPGA fabric\001
+4 1 0 40 -1 18 18 0.0000 4 270 750 6705 2250 pipe0\001
+4 1 0 40 -1 18 18 0.0000 4 270 750 6705 4050 pipe1\001
+4 1 0 50 -1 18 24 0.0000 4 390 1245 15525 6210 config\001
