+++ doc/spif_config/spin_per_if.fig
@@ -143,5 +143,5 @@ Single
 4 1 20 40 -1 18 20 0.0000 4 240 1380 7200 6900 OUTPUT\001
 4 1 20 40 -1 18 20 0.0000 4 315 1905 12300 975 SpiNNTools\001
 4 1 20 40 -1 18 20 0.0000 4 315 1680 12375 8250 SpiNNaker\001
-4 1 20 40 -1 18 20 0.0000 4 315 2700 7200 8250 SpiNNaker FPGA\001
+4 1 20 40 -1 18 20 0.0000 4 315 1530 7200 8250 spiNNlink\001
 4 1 24 40 -1 18 20 0.0000 4 315 585 3525 8250 spif\001
